requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
torch>=1.10.0
gymnasium>=0.28.1
tqdm>=4.62.0
scikit-learn>=1.0.0
alpaca-py>=0.8.0
